transitioned to a mix of dapper and EFcore. added orders, added products and customers.

+++ ecommerce.Core/Services/Interfaces/IService.cs
@@ -6,7 +6,7 @@ namespace ecommerce.Core.Services.Interfaces
     public interface IService<TEntity>
     {
         Task<IEnumerable<TEntity>> GetAllAsync();
-        Task<TEntity> GetByIdAsync(long id);
+        Task<TEntity> GetByIdAsync(int id);
         Task AddAsync(TEntity entity);
         Task UpdateAsync(TEntity entity);
         Task DeleteAsync(TEntity entity);
